requirements.txt
flask-socketio
requests
characterai==0.8.0
typing_extensions
websockets
python-dotenv
